wrote createBookmark.php
no idea if it works cause idk how to test it yet but this is most of the code that will be needed for creating bookmarks. still need to get the media ID and userID from somewhere and need to add a dropdown for watch status which makes the rating and review options go away.